# managed/gateway/openapi/managed.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Steel Managed Operations API",
    "version": "2026-07-01",
    "license": {
      "name": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/v1/managed/workers": {
      "get": {
        "operationId": "listManagedWorkers",
        "tags": [
          "Managed operations"
        ],
        "security": [
          {
            "cloudflareAccess": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]+$",
              "minLength": 2
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "minItems": 1,
              "items": {
                "type": "string",
                "enum": [
                  "DISCOVERED",
                  "REACHABLE",
                  "IDLE",
                  "RESERVED",
                  "STARTING",
                  "LIVE",
                  "RELEASING",
                  "UNREACHABLE",
                  "QUARANTINED",
                  "DRAINING"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful managed operation",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerList"
                }
              }
            }
          },
          "400": {
            "description": "Managed error (400)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "INVALID_ARGUMENT",
                                "INVALID_CURSOR"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Managed error (401)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "ACCESS_AUTH_REQUIRED"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Managed error (403)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "ACCESS_FORBIDDEN"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Managed error (409)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "LIST_CURSOR_RESTARTED"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "410": {
            "description": "Managed error (410)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "LIST_CURSOR_EXPIRED"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "502": {
            "description": "Managed error (502)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "UPSTREAM_BAD_RESPONSE"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Managed error (503)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "MANAGED_SNAPSHOT_CAPACITY"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/managed/sessions": {
      "get": {
        "operationId": "listManagedSessions",
        "tags": [
          "Managed operations"
        ],
        "security": [
          {
            "cloudflareAccess": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]+$",
              "minLength": 2
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "minItems": 1,
              "items": {
                "type": "string",
                "enum": [
                  "QUEUED",
                  "STARTING",
                  "LIVE",
                  "RELEASING",
                  "RELEASED",
                  "FAILED",
                  "LOST"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful managed operation",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionList"
                }
              }
            }
          },
          "400": {
            "description": "Managed error (400)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "INVALID_ARGUMENT",
                                "INVALID_CURSOR"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Managed error (401)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "ACCESS_AUTH_REQUIRED"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Managed error (409)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "LIST_CURSOR_RESTARTED"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "410": {
            "description": "Managed error (410)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "LIST_CURSOR_EXPIRED"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "502": {
            "description": "Managed error (502)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "UPSTREAM_BAD_RESPONSE"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Managed error (503)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "MANAGED_SNAPSHOT_CAPACITY"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cloudflareAccess": {
        "type": "apiKey",
        "in": "header",
        "name": "Cf-Access-Jwt-Assertion"
      }
    },
    "schemas": {
...
      "SessionList": {
        "type": "object",
        "properties": {
          "apiVersion": {
            "type": "string",
            "enum": [
              "2026-07-01"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "sessionId": {
                  "type": "string",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                },
                "state": {
                  "type": "string",
                  "enum": [
                    "QUEUED",
                    "STARTING",
                    "LIVE",
                    "RELEASING",
                    "RELEASED",
                    "FAILED",
                    "LOST"
                  ]
                },
                "workerId": {
                  "type": "string",
                  "pattern": "^worker-(?:0[0-9]|[12][0-9]|3[01])$"
                },
                "instanceId": {
                  "type": "string",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                },
                "admissionId": {
                  "type": "string",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "startedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "endedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "failureCode": {
                  "type": "string",
                  "pattern": "^[A-Z][A-Z0-9_]{0,63}$"
                }
              },
              "required": [
                "sessionId",
                "state",
                "createdAt"
              ],
              "additionalProperties": false
            }
          },
          "page": {
            "type": "object",
            "properties": {
              "pageSize": {
                "type": "integer",
                "minimum": 1,
                "maximum": 100
              },
              "nextCursor": {
                "type": "string",
                "pattern": "^[A-Za-z0-9_-]+$",
                "minLength": 2
              },
              "snapshotCursor": {
                "type": "string",
                "pattern": "^[A-Za-z0-9_-]+$",
                "minLength": 2
              },
              "hasMore": {
                "type": "boolean"
              }
            },
            "required": [
              "pageSize",
              "snapshotCursor",
              "hasMore"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "apiVersion",
          "items",
          "page"
        ],
        "additionalProperties": false
      },
...
      "WorkerList": {
        "type": "object",
        "properties": {
          "apiVersion": {
            "type": "string",
            "enum": [
              "2026-07-01"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "workerId": {
                  "type": "string",
                  "pattern": "^worker-(?:0[0-9]|[12][0-9]|3[01])$"
                },
                "instanceId": {
                  "type": "string",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                },
                "state": {
                  "type": "string",
                  "enum": [
                    "DISCOVERED",
                    "REACHABLE",
                    "IDLE",
                    "RESERVED",
                    "STARTING",
                    "LIVE",
                    "RELEASING",
                    "UNREACHABLE",
                    "QUARANTINED",
                    "DRAINING"
                  ]
                },
                "sessionId": {
                  "type": "string",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                },
                "lastSeenAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "stateChangedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "failureCode": {
                  "type": "string",
                  "pattern": "^[A-Z][A-Z0-9_]{0,63}$"
                }
              },
              "required": [
                "workerId",
                "instanceId",
                "state",
                "lastSeenAt",
                "stateChangedAt"
              ],
              "additionalProperties": false
            }
          },
          "page": {
            "type": "object",
            "properties": {
              "pageSize": {
                "type": "integer",
                "minimum": 1,
                "maximum": 100
              },
              "nextCursor": {
                "type": "string",
                "pattern": "^[A-Za-z0-9_-]+$",
                "minLength": 2
              },
              "snapshotCursor": {
                "type": "string",
                "pattern": "^[A-Za-z0-9_-]+$",
                "minLength": 2
              },
              "hasMore": {
                "type": "boolean"
              }
            },
            "required": [
              "pageSize",
              "snapshotCursor",
              "hasMore"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "apiVersion",
          "items",
          "page"
        ],
        "additionalProperties": false
      },
      "JsonValue": {
        "oneOf": [
          {
            "enum": [
              null
            ]
          },
          {
            "type": "boolean"
          },
          {
            "type": "number"
          },
          {
            "type": "string"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            }
          },
          {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            }
          }
        ]
      },
      "ManagedErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "apiVersion",
          "error"
        ],
        "properties": {
          "apiVersion": {
            "type": "string",
            "enum": [
              "2026-07-01"
            ]
          },
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "retryable",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "INVALID_ARGUMENT",
                  "INVALID_CURSOR",
                  "ACCESS_AUTH_REQUIRED",
                  "ACCESS_FORBIDDEN",
                  "SESSION_NOT_FOUND",
                  "ADMISSION_NOT_FOUND",
                  "RESULT_NOT_FOUND",
                  "TOOL_NOT_FOUND",
                  "ROUTE_NOT_FOUND",
                  "METHOD_NOT_ALLOWED",
                  "NOT_ACCEPTABLE",
                  "MANAGED_SESSION_REQUIRED",
                  "SESSION_STATE_CONFLICT",
                  "CREATE_TOKEN_CONFLICT",
                  "EVENT_CURSOR_RESTARTED",
                  "LIST_CURSOR_RESTARTED",
                  "INSTANCE_LOST",
                  "EVENT_CURSOR_EXPIRED",
                  "LIST_CURSOR_EXPIRED",
                  "BODY_TOO_LARGE",
                  "RESULT_TOO_LARGE",
                  "UNSUPPORTED_MEDIA_TYPE",
                  "MANAGED_FEATURE_UNSUPPORTED",
                  "TOOL_INPUT_INVALID",
                  "ADMISSION_TIMEOUT",
                  "QUEUE_FULL",
                  "RATE_LIMITED",
                  "UPSTREAM_BAD_RESPONSE",
                  "NO_REACHABLE_WORKER",
                  "MANAGER_DRAINING",
                  "MANAGED_LIMITER_CAPACITY",
                  "MANAGED_TICKET_CAPACITY",
                  "MANAGED_IDEMPOTENCY_CAPACITY",
                  "MANAGED_SNAPSHOT_CAPACITY",
                  "MANAGED_INGRESS_CAPACITY",
                  "MANAGED_AI_RESULT_CAPACITY",
                  "MANAGED_ACTION_CAPACITY",
                  "MANAGED_WS_CAPACITY",
                  "WORKER_JOURNAL_CAPACITY",
                  "UPSTREAM_TIMEOUT",
                  "TOOL_TIMEOUT"
                ]
              },
              "message": {
                "type": "string",
                "minLength": 1,
                "maxLength": 1024
              },
              "retryable": {
                "type": "boolean"
              },
              "requestId": {
                "type": "string",
                "format": "uuid"
              },
              "details": {
                "$ref": "#/components/schemas/JsonValue"
              }
            }
          }
        }
      }
    }
  }
}